Small improvements and added icon to decide when to create the selection object

+++ src/dataModel/DatasetDTO.h
@@ -27,10 +27,10 @@ class DatasetDTO : public QObject
     Q_PROPERTY(QString image_alignment_id READ imageAlignmentId WRITE imageAlignmentId)
     Q_PROPERTY(QString tissue READ statTissue WRITE statTissue)
     Q_PROPERTY(QString species READ statSpecies WRITE statSpecies)
-    Q_PROPERTY(unsigned overall_feature_count READ statBarcodes WRITE statBarcodes)
-    Q_PROPERTY(unsigned overall_hit_count READ statGenes WRITE statGenes)
-    Q_PROPERTY(unsigned unique_barcode_count READ statUniqueBarcodes WRITE statUniqueBarcodes)
-    Q_PROPERTY(unsigned unique_gene_count READ statUniqueGenes WRITE statUniqueGenes)
+    Q_PROPERTY(int overall_feature_count READ statBarcodes WRITE statBarcodes)
+    Q_PROPERTY(int overall_hit_count READ statGenes WRITE statGenes)
+    Q_PROPERTY(int unique_barcode_count READ statUniqueBarcodes WRITE statUniqueBarcodes)
+    Q_PROPERTY(int unique_gene_count READ statUniqueGenes WRITE statUniqueGenes)
     Q_PROPERTY(QVariantList overall_hit_quartiles READ hitsQuartiles WRITE hitsQuartiles)
     Q_PROPERTY(QVariantList gene_pooled_hit_quartiles READ hitsPooledQuartiles WRITE
                    hitsPooledQuartiles)
@@ -59,13 +59,13 @@ public:
     void imageAlignmentId(const QString &alignmentId) { m_dataset.imageAlignmentId(alignmentId); }
     void statTissue(const QString &tissue) { m_dataset.statTissue(tissue); }
     void statSpecies(const QString &species) { m_dataset.statSpecies(species); }
-    void statBarcodes(unsigned barcodes) { m_dataset.statBarcodes(barcodes); }
-    void statGenes(unsigned genes) { m_dataset.statGenes(genes); }
-    void statUniqueBarcodes(unsigned unique_barcodes)
+    void statBarcodes(int barcodes) { m_dataset.statBarcodes(barcodes); }
+    void statGenes(int genes) { m_dataset.statGenes(genes); }
+    void statUniqueBarcodes(int unique_barcodes)
     {
         m_dataset.statUniqueBarcodes(unique_barcodes);
     }
-    void statUniqueGenes(unsigned unique_genes) { m_dataset.statUniqueGenes(unique_genes); }
+    void statUniqueGenes(int unique_genes) { m_dataset.statUniqueGenes(unique_genes); }
     void hitsQuartiles(const QVariantList &hitQuartiles)
     {
         m_dataset.hitsQuartiles(unserializeVector<float>(hitQuartiles));
@@ -90,10 +90,10 @@ public:
     const QString imageAlignmentId() const { return m_dataset.imageAlignmentId(); }
     const QString statTissue() const { return m_dataset.statTissue(); }
     const QString statSpecies() const { return m_dataset.statSpecies(); }
-    unsigned statBarcodes() const { return m_dataset.statBarcodes(); }
-    unsigned statGenes() const { return m_dataset.statGenes(); }
-    unsigned statUniqueBarcodes() const { return m_dataset.statUniqueBarcodes(); }
-    unsigned statUniqueGenes() const { return m_dataset.statUniqueGenes(); }
+    int statBarcodes() const { return m_dataset.statBarcodes(); }
+    int statGenes() const { return m_dataset.statGenes(); }
+    int statUniqueBarcodes() const { return m_dataset.statUniqueBarcodes(); }
+    int statUniqueGenes() const { return m_dataset.statUniqueGenes(); }
     const QVariantList hitsQuartiles() const
     {
         return serializeVector<float>(m_dataset.hitsQuartiles());
@@ -112,9 +112,8 @@ public:
     const QString created() const { return m_dataset.created(); }
     const QString lastModified() const { return m_dataset.lastModified(); }
 
-    // toJson is needed to send PUT/POST requests as the JSON content of the
-    // object
-    // is appended to the request
+    // toJson is needed to send PUT/POST requests as the JSON representation of the
+    // object is appended to the request
     // TODO transform this to obtain fields dynamically using the meta_properties
     QByteArray toJson() const
     {
@@ -124,10 +123,10 @@ public:
         jsonObj["image_alignment_id"] = imageAlignmentId();
         jsonObj["tissue"] = statTissue();
         jsonObj["species"] = statSpecies();
-        jsonObj["overall_feature_count"] = static_cast<int>(statBarcodes());
-        jsonObj["overall_hit_count"] = static_cast<int>(statGenes());
-        jsonObj["unique_barcode_count"] = static_cast<int>(statUniqueBarcodes());
-        jsonObj["unique_gene_count"] = static_cast<int>(statUniqueGenes());
+        jsonObj["overall_feature_count"] = statBarcodes();
+        jsonObj["overall_hit_count"] = statGenes();
+        jsonObj["unique_barcode_count"] = statUniqueBarcodes();
+        jsonObj["unique_gene_count"] = statUniqueGenes();
         QJsonArray hitsQuartiles;
         for (const auto &item : m_dataset.hitsQuartiles()) {
             hitsQuartiles.append(item);
